Update 4
Added functionality to computer player move where it will make a move automatically if it's a win.  Then it checks the real player moves and if there's any positions where the player can win it will block it.  Created a function for the player move where it will take in an index and make a move at that location.  Created errors to allow only open position inputs.